Move vdom, hole, block, parser, into view
These types are all interdependent upon each other. Easier to have them
in a single module.
