+++ typescript/packages/common-html/src/index.ts
@@ -1,8 +1,6 @@
-export { view, View, Context } from "./view.js";
+export { view, View, Context, VNode, Var } from "./view.js";
 export { html } from "./html.js";
 export { render, setNodeSanitizer, setEventSanitizer } from "./render.js";
-export { VNode } from "./vnode.js";
-export { Hole } from "./hole.js";
 export { Reactive } from "./reactive.js";
 export { setDebug } from "./logger.js";
 export { cancel, Cancel } from "./cancel.js";
