+++ typescript/packages/common-html/src/test/html.test.ts
@@ -1,6 +1,6 @@
 import * as assert from "node:assert";
 import html from "../html.js";
-import * as hole from "../hole.js";
+import { isVar } from "../view.js";
 import { state, stream } from "../state.js";
 
 describe("html", () => {
@@ -15,9 +15,9 @@ describe("html", () => {
     `;
 
     // @ts-ignore - ignore for test
-    assert.strict(hole.isHole(view.template.children[0].props.onclick));
+    assert.strict(isVar(view.template.children[0].props.onclick));
 
     // @ts-ignore - ignore for test
-    assert.strict(hole.isHole(view.template.children[0].children[0]));
+    assert.strict(isVar(view.template.children[0].children[0]));
   });
 });
